+++ src/bin/pica/config.rs
@@ -7,9 +7,22 @@ use serde::{Deserialize, Serialize};
 
 use crate::commands::*;
 
+#[derive(Debug, Clone, Copy, Serialize, Deserialize)]
+pub(crate) enum NormalizationForm {
+    #[serde(alias = "nfc")]
+    Nfc,
+    #[serde(alias = "nfkc")]
+    Nfkc,
+    #[serde(alias = "nfd")]
+    Nfd,
+    #[serde(alias = "nfkd")]
+    Nfkd,
+}
+
 #[derive(Debug, Default, Deserialize, Serialize)]
 #[serde(rename_all = "kebab-case")]
 pub(crate) struct GlobalConfig {
+    pub(crate) translit: Option<NormalizationForm>,
     pub(crate) skip_invalid: Option<bool>,
 }