+++ src/bin/pica/commands/filter.rs
@@ -1,25 +1,19 @@
 use std::ffi::OsString;
 use std::fs::read_to_string;
-use std::io::{self, Read};
 use std::path::PathBuf;
-use std::str::FromStr;
 
 use clap::{value_parser, Parser};
-use lazy_static::lazy_static;
-use pica::matcher::{
-    MatcherFlags, OccurrenceMatcher, RecordMatcher, TagMatcher,
+use pica_matcher::{
+    MatcherOptions, OccurrenceMatcher, ParseMatcherError,
+    RecordMatcher, TagMatcher,
 };
-use pica::{Path, PicaWriter, Reader, ReaderBuilder, WriterBuilder};
+use pica_record::io::{ReaderBuilder, RecordsIterator, WriterBuilder};
 use serde::{Deserialize, Serialize};
 
 use crate::common::FilterList;
 use crate::translit::translit_maybe2;
 use crate::util::{CliError, CliResult};
 use crate::{gzip_flag, skip_invalid_flag, Config};
-
-lazy_static! {
-    static ref IDN_PATH: Path = Path::from_str("003@.0").unwrap();
-}
 
 #[derive(Debug, Deserialize, Serialize)]
 #[serde(rename_all = "kebab-case")]
@@ -28,9 +22,10 @@ pub(crate) struct FilterConfig {
     pub(crate) gzip: Option<bool>,
 }
 
+/// Filter records by whether the given filter expression matches
 #[derive(Parser, Debug)]
 pub(crate) struct Filter {
-    /// Skip invalid records that can't be decoded
+    /// Skip invalid records that can't be decoded as normalized PICA+
     #[arg(short, long)]
     skip_invalid: bool,
 
@@ -49,42 +44,67 @@ pub(crate) struct Filter {
         default_value = "75")]
     strsim_threshold: u8,
 
-    /// Reduce the record to fields which are specified in <REDUCE>
-    #[arg(long, short = 'R', default_value = "")]
-    reduce: String,
+    /// Keep only fields specified by a list of predicates.
+    #[arg(long, short)]
+    keep: Option<String>,
+
+    /// Discard fields specified by a list of predicates.
+    #[arg(long, short)]
+    discard: Option<String>,
 
     /// Take filter expressions from <EXPR_FILE>
+    ///
+    /// Note: Using a expression file still requires a filter; e.g
+    /// `003@.0?`.
     #[arg(long = "file", short = 'f')]
     expr_file: Option<PathBuf>,
 
     /// Ignore records which are *not* explicitly listed in one of the
-    /// given allow-lists. An allow-list must be an CSV, whereby the
-    /// first column contains the IDN (003@.0).
+    /// given allow-lists.
+    ///
+    /// An allow-list must be an CSV, whereby the first column contains
+    /// the IDN (003@.0) or an Apache Arrow file with an `idn` column.
+    /// If the file extension is `.feather`, `.arrow`, or `.ipc` the
+    /// file is automatically interpreted as Apachae Arrow;
+    /// otherwise the file is read as CSV.
     #[arg(long, short = 'A')]
     allow_list: Vec<PathBuf>,
 
     /// Ignore records which are explicitly listed in one of the
-    /// given deny-lists. An deny-list must be an CSV, whereby the
-    /// first column contains the IDN (003@.0).
+    /// given deny-lists.
+    ///
+    /// An allow-list must be an CSV, whereby the first column contains
+    /// the IDN (003@.0) or an Apache Arrow file with an `idn` column.
+    /// If the file extension is `.feather`, `.arrow`, or `.ipc` the
+    /// file is automatically interpreted as Apachae Arrow;
+    /// otherwise the file is read as CSV.
     #[arg(long, short = 'D')]
     deny_list: Vec<PathBuf>,
 
     /// Limit the result to first <n> records
+    ///
+    /// Note: A limit value `0` means no limit.
     #[arg(long, short, value_name = "n", default_value = "0")]
     limit: usize,
 
     /// Connects the filter with additional expressions using the
     /// logical AND-operator (conjunction)
+    ///
+    /// This option can't be combined with `--or` or `--not`.
     #[arg(long, conflicts_with_all = ["or", "not"])]
     and: Vec<String>,
 
     /// Connects the filter with additional expressions using the
     /// logical OR-operator (disjunction)
+    ///
+    /// This option can't be combined with `--and` or `--not`.
     #[arg(long, conflicts_with_all = ["and", "not"])]
     or: Vec<String>,
 
     /// Connects the filter with additional expressions using the
     /// logical NOT-operator (negation)
+    ///
+    /// This option can't be combined with `--and` or `--or`.
     #[arg(long, conflicts_with_all = ["and", "or"])]
     not: Vec<String>,
 
@@ -93,7 +113,9 @@ pub(crate) struct Filter {
     gzip: bool,
 
     /// Append to the given file, do not overwrite
-    #[arg(long)]
+    ///
+    /// Warning: This option can't be used when writing to a gzip file.
+    #[arg(long, conflicts_with = "gzip")]
     append: bool,
 
     /// Write simultaneously to the file <filename> and stdout
@@ -107,14 +129,16 @@ pub(crate) struct Filter {
     /// A filter expression used for searching
     filter: String,
 
-    /// Read one or more files in normalized PICA+ format.
+    /// Read one or more files in normalized PICA+ format
+    ///
+    /// If no filenames where given or a filename is "-", data is read
+    /// from standard input (stdin).
     #[arg(default_value = "-", hide_default_value = true)]
     filenames: Vec<OsString>,
 }
 
 impl Filter {
     pub(crate) fn run(self, config: &Config) -> CliResult<()> {
-        // let limit = self.limit.unwrap_or_default();
         let gzip_compression = gzip_flag!(self.gzip, config.filter);
         let skip_invalid = skip_invalid_flag!(
             self.skip_invalid,
@@ -122,7 +146,7 @@ impl Filter {
             config.global
         );
 
-        let mut writer: Box<dyn PicaWriter> = WriterBuilder::new()
+        let mut writer = WriterBuilder::new()
             .gzip(gzip_compression)
             .append(self.append)
             .from_path_or_stdout(self.output)?;
@@ -137,30 +161,10 @@ impl Filter {
             None => None,
         };
 
-        let items = self
-            .reduce
-            .split(',')
-            .map(str::trim)
-            .filter(|s| !s.is_empty());
-
-        let mut reducers = vec![];
-        for item in items {
-            let (tag, occ) = if let Some(pos) = item.rfind('/') {
-                (&item[0..pos], &item[pos..])
-            } else {
-                (item, "/*")
-            };
-
-            let tag = TagMatcher::new(tag).map_err(|_| {
-                CliError::Other("invalid reduce value".to_string())
-            })?;
-
-            let occ = OccurrenceMatcher::new(occ).map_err(|_| {
-                CliError::Other("invalid reduce value".to_string())
-            })?;
-
-            reducers.push((tag, occ));
-        }
+        let discard_predicates =
+            parse_predicates(&self.discard.unwrap_or_default())?;
+        let keep_predicates =
+            parse_predicates(&self.keep.unwrap_or_default())?;
 
         let filter_str = if let Some(filename) = self.expr_file {
             read_to_string(filename).unwrap()
@@ -184,38 +188,20 @@ impl Filter {
         };
 
         if !self.and.is_empty() {
-            let predicates = self
-                .and
-                .iter()
-                .map(RecordMatcher::new)
-                .collect::<Result<Vec<_>, _>>()?;
-
-            for expr in predicates.into_iter() {
-                filter = filter & expr;
+            for predicate in self.and.iter() {
+                filter = filter & RecordMatcher::new(predicate)?;
             }
         }
 
         if !self.not.is_empty() {
-            let predicates = self
-                .not
-                .iter()
-                .map(RecordMatcher::new)
-                .collect::<Result<Vec<_>, _>>()?;
-
-            for expr in predicates.into_iter() {
-                filter = filter & !expr;
+            for predicate in self.not.iter() {
+                filter = filter & !RecordMatcher::new(predicate)?;
             }
         }
 
         if !self.or.is_empty() {
-            let predicates = self
-                .or
-                .iter()
-                .map(RecordMatcher::new)
-                .collect::<Result<Vec<_>, _>>()?;
-
-            for expr in predicates.into_iter() {
-                filter = filter | expr;
+            for predicate in self.or.iter() {
+                filter = filter | RecordMatcher::new(predicate)?;
             }
         }
 
@@ -232,65 +218,85 @@ impl Filter {
         };
 
         let mut count = 0;
-        let flags = MatcherFlags {
-            strsim_threshold: self.strsim_threshold as f64 / 100.0,
-            ignore_case: self.ignore_case,
-        };
+        let options = MatcherOptions::new()
+            .strsim_threshold(self.strsim_threshold as f64 / 100.0)
+            .case_ignore(self.ignore_case);
 
         for filename in self.filenames {
-            let builder =
-                ReaderBuilder::new().skip_invalid(skip_invalid);
-            let mut reader: Reader<Box<dyn Read>> = match filename
-                .to_str()
-            {
-                Some("-") => builder.from_reader(Box::new(io::stdin())),
-                _ => builder.from_path(filename)?,
-            };
+            let mut reader =
+                ReaderBuilder::new().from_path(filename)?;
 
-            for result in reader.byte_records() {
-                let mut record = result?;
-                let idn = record.path(&IDN_PATH);
-                let idn = idn.first();
-
-                if !allow_list.is_empty() {
-                    if let Some(idn) = idn {
-                        if !allow_list.contains(*idn) {
+            while let Some(result) = reader.next() {
+                match result {
+                    Err(e) => {
+                        if e.is_invalid_record() && skip_invalid {
                             continue;
-                        }
-                    } else {
-                        continue;
-                    }
-                }
-
-                if !deny_list.is_empty() {
-                    if let Some(idn) = idn {
-                        if deny_list.contains(*idn) {
-                            continue;
+                        } else {
+                            return Err(e.into());
                         }
                     }
-                }
+                    Ok(mut record) => {
+                        if !allow_list.is_empty()
+                            && !allow_list.check(&record)
+                        {
+                            continue;
+                        }
 
-                let mut is_match = filter.is_match(&record, &flags);
-                if self.invert_match {
-                    is_match = !is_match;
-                }
+                        if !deny_list.is_empty()
+                            && deny_list.check(&record)
+                        {
+                            continue;
+                        }
 
-                if is_match {
-                    if !reducers.is_empty() {
-                        record.reduce(&reducers);
+                        let mut is_match =
+                            filter.is_match(&record, &options);
+
+                        if self.invert_match {
+                            is_match = !is_match;
+                        }
+
+                        if is_match {
+                            if !keep_predicates.is_empty() {
+                                record.retain(|field| {
+                                    for (t, o) in keep_predicates.iter()
+                                    {
+                                        if t.is_match(field.tag())
+                                            && *o == field.occurrence()
+                                        {
+                                            return true;
+                                        }
+                                    }
+                                    false
+                                });
+                            }
+
+                            if !discard_predicates.is_empty() {
+                                record.retain(|field| {
+                                    for (t, o) in
+                                        discard_predicates.iter()
+                                    {
+                                        if t.is_match(field.tag())
+                                            && *o == field.occurrence()
+                                        {
+                                            return false;
+                                        }
+                                    }
+                                    true
+                                });
+                            }
+
+                            writer.write_byte_record(&record)?;
+                            if let Some(ref mut writer) = tee_writer {
+                                writer.write_byte_record(&record)?;
+                            }
+
+                            count += 1;
+                        }
+
+                        if self.limit > 0 && count >= self.limit {
+                            break;
+                        }
                     }
-
-                    writer.write_byte_record(&record)?;
-
-                    if let Some(ref mut writer) = tee_writer {
-                        writer.write_byte_record(&record)?;
-                    }
-
-                    count += 1;
-                }
-
-                if self.limit > 0 && count >= self.limit {
-                    break;
                 }
             }
         }
@@ -302,4 +308,27 @@ impl Filter {
 
         Ok(())
     }
+}
+
+fn parse_predicates(
+    s: &str,
+) -> Result<Vec<(TagMatcher, OccurrenceMatcher)>, ParseMatcherError> {
+    let items = s.split(',').map(str::trim).filter(|s| !s.is_empty());
+    let mut result = vec![];
+
+    for item in items {
+        if let Some(pos) = item.rfind('/') {
+            result.push((
+                TagMatcher::new(&item[0..pos])?,
+                OccurrenceMatcher::new(&item[pos..])?,
+            ));
+        } else {
+            result.push((
+                TagMatcher::new(item)?,
+                OccurrenceMatcher::None,
+            ));
+        }
+    }
+
+    Ok(result)
 }