+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * #572 Rephrase misleading documentation (`invalid`)
 * #590 Stabilize `slice` command
 * #592 Transliteration of matcher expressions
+* #564 Add `--seed` option to `sample` command
 
 ## 0.14.1